feat: add serverController with REST endpoints

+++ src/server/server.service.ts
@@ -1,18 +1,35 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  ForbiddenException,
+} from '@nestjs/common';
 import { CreateServerDto } from './dto/create-server.dto';
 import { UpdateServerDto } from './dto/update-server.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Server } from './entities/server.entity';
 import { Repository } from 'typeorm';
-
+import { AuthenticatedRequest } from '../auth/interfaces/authenticated-request.interface';
 @Injectable()
 export class ServerService {
   constructor(
     @InjectRepository(Server)
     private readonly serverRepository: Repository<Server>,
   ) {}
-  async create(createServerDto: CreateServerDto) {
-    return await this.serverRepository.save(createServerDto);
+
+  async create(createServerDto: CreateServerDto, req: AuthenticatedRequest) {
+    const server = this.serverRepository.create({
+      ...createServerDto,
+      ownerId: req.user.userId,
+    });
+
+    const savedServer = await this.serverRepository.save(server);
+
+    return {
+      message: 'Server created successfully',
+      data: {
+        server: savedServer,
+      },
+    };
   }
 
   async findAll() {
@@ -40,13 +57,19 @@ export class ServerService {
     };
   }
 
-  async update(id: number, updateServerDto: UpdateServerDto) {
+  async update(
+    id: number,
+    updateServerDto: UpdateServerDto,
+    req: AuthenticatedRequest,
+  ) {
     const server = await this.serverRepository.findOne({
       where: { id },
     });
 
     if (!server) throw new NotFoundException(`Server with ID ${id} not found`);
-
+    if (server.ownerId !== req.user.userId) {
+      throw new ForbiddenException('Only server owner can update the server');
+    }
     Object.assign(server, updateServerDto);
     const updatedUser = await this.serverRepository.save(server);
     return {
@@ -57,13 +80,15 @@ export class ServerService {
     };
   }
 
-  async remove(id: number) {
+  async remove(id: number, req: AuthenticatedRequest) {
     const server = await this.serverRepository.findOne({
       where: { id },
     });
 
     if (!server) throw new NotFoundException(`Server with ID ${id} not found`);
-
+    if (server.ownerId !== req.user.userId) {
+      throw new ForbiddenException('Only server owner can update the server');
+    }
     await this.serverRepository.remove(server);
     return {
       messgae: `Server with ID ${id} removed`,
